refactor : 좌석 예매 -> 예매 완료 화면에 티켓 전달시 좌석 리스트를 전달하도록 수정
