+++ app/src/main/java/woowacourse/movie/domain/model/ticket/Ticket.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.domain.model.ticket
 
 import woowacourse.movie.domain.model.booking.Booking
 import woowacourse.movie.domain.model.booking.PeopleCount
+import woowacourse.movie.domain.model.seat.Seat
 import java.io.Serializable
 import java.time.LocalDate
 import java.time.LocalTime
@@ -11,12 +12,13 @@ data class Ticket(
     val bookingDate: LocalDate,
     val bookingTime: LocalTime,
     val count: PeopleCount,
-    val seats: String,
+    val seats: Set<Seat>,
     val price: Int,
 ) : Serializable {
     companion object {
         fun initialize(
             booking: Booking,
+            seats: Set<Seat>,
             price: Int,
         ): Ticket {
             return Ticket(
@@ -24,7 +26,7 @@ data class Ticket(
                 bookingDate = booking.bookingDate,
                 bookingTime = booking.bookingTime,
                 count = booking.count,
-                seats = "",
+                seats = seats,
                 price = price,
             )
         }
